Move Algebra into Control.Algebra.Internal.
